This is the more stable bootloader from https://github.com/dabit20/R9MM_bootloader/releases/tag/v1.0
The original bootloader tends to destroy the application data in the MCU flash when the flightcontroller is already active before the R9MM is powered. This can happen when the FC is plugged into USB before a battery is connected.

This bootloader improves on that. It is stil not perfect since it still tends to enter bootloader state instead of jumping to the main application during startup, but that condition can be resolved by replugging the battery; the flash contents stay intact.