Scale RSSI outputs for BetaFlight and OpenTX
Neither BetaFlight nor OpenTX currently understand the RSSI values
that we are sending. This PR allows both consumers to display
consistent RSSI dB values in the negative range. The use of 8 bit
values in various places limits the lowest RSSI that can be shown
to -128 dB and the upper limit to 0 dB.

Signed-off-by: JBKingdon <james.kingdon@gmail.com>