Sequential and hybrid switch modes
Sequential switches sends 1 switch state per RC packet using a 3 bit switch index
and a 2 bit switch value. This supports 8 3-position switches. The transmitter
keeps an array of 'sent' switch values and compares this with the current values
in order to detect switches that have changed value. The lowest index changed
switch is sent to the receiver. If no switches have changed then each switch is
sent in round robin fashion.

For sequential switches the last analog channel is reduced to 10 bit resolution.

Hybrid switch mode sends the lowest index switch with every packet so as to
ensure the lowest possible latency. This switch is intended to be used for
arming. The remaining switches are sent in the same way as described above for
sequential switches. All 8 switches support 3-position output.

For hybrid switches all analog channels are reduced to 10 bit resolution.

Signed-off-by: JBKingdon <james.kingdon@gmail.com>

Conflicts:
	src/lib/CRSF/CRSF.cpp
	src/lib/CRSF/CRSF.h
	src/src/tx_main.cpp