Add fallback to resync if wait for next confirmation is stuck
If the stubborn receiver resets right after the last part of a message is received all further communication would be stuck:
The sender would send package index 0 to get the final confirmation and the receiver would wait for package index 1 to start the first message.

This change makes sure that the same timeout that is used during the transmission of data is also applied for the final confirmation.
Switching to the RESYNC state forces the receiver in the correct state and allows to resume communication.