Fix issue "RX Serial not found", missing bootloader
I hit an issue on mnibus f4, the ExpressLRS-Configurator
can't flash the elrs binary due to not able to detect bootloader.
The issue is cause by in correct parsing of the delimiter
when a line is broken into two separate serial reads.

This code in question is:
	r = buf + data[:offset]
It should be:
	r = (buf + data)[:offset]

I also notice duplicated code to handle one read vs more
than one read of a line. I clean up and unify the duplicated
code.

With this change, I verify the elrs firmware can be upload correctly.