Use modelId for configuration and model matching (#765)
* Make the switch mode configurable from LUA

Removes the switch and telemetry defines

* Add modelId parsing into CRSF

Use radio modelId as an index for different configs
i.e. the rate, tlm ratio, max power and switch mode are per model

* Implement model match on the TX, with updates to LUA script

* Add model match to RX (specified by a user define)

* RX webpage update to allow setting model match id

* Set model match from command line

Introduce a telemetry command so the model match can be updated via passthrough

* Add MSP command to set model match

This enables us to have a LUA command that sends a MSP message to a connected receiver to set its model match id.