Simplify setting RX model match
When you enable model match and are connected to an RX the RX store the model number and reboots, the TX starts sending RF with the model match enabled.
To remove model matching, connect to the model matched RX and disable model match in the LUA, the RX will set model matching off and reboot, the TX starts sending RF with model match disabled.