Adds missing vario tlm sensor and a generic way to handle custom telemetry frame types
iNav uses a vario tlm sensor that only has a 2 byte payload.

KISS uses a custom frame for requests so this ID is handled like a MSP request.

CRSF defines custom frame types for different FCs. Currently known are frames for KISS and Ardupilot. Until now only MSP was supported but this commit changes the 64 byte buffer for a MSP response to a general fallback buffer.

That buffer is used for all frame types that are not explicitly added to the telemetry sensor list.
Since a rx isn't connected to a KISS FC and Betaflight FC at the same time this should not be an issue that the buffer is used for different custom frame types at the same time.