Adds missing vario tlm sensor and a generic tlm response
Adds a vario sensor used by iNav

CRSF defines custom frame types for different FCs. Currently known are frames for KISS and Ardupilot.
Until now only MSP was supported but this commit adds 2x a 64 byte buffer for a general tlm response from other firmware projects.

That buffer is used for all frame types that are not explicitly added to the telemetry sensor list. If the first slot is used there is a second slot that can be used.