Add rx device info frame as response to pings (#1021)
CRSF defines a message called CRSF_FRAMETYPE_DEVICE_PING. This message allows to ping all devices (also called broadcast) or a certain device. If a device receives a ping it should respond with a CRSF_FRAMETYPE_DEVICE_INFO message.

The tx module already responds to this message for the lua script support. With #997 our lua script also shows other devices but the rx did not get listed since it did not respond to the ping message.

This PR adds the device name for the rx targets and adds two responses to pings:

ping received from tx module (from the lua script): reply OTA
ping received from FC (needed for ardupilot): reply via UART

The method to generate the device info frame is moved in the crsf class and both tx+rx use the same method.
To inject the device info frame in the usual telemetry buffer a change in the logic was needed. The device info response of the FC would overwrite the device info response of the rx so the telemetry class makes sure that different buffers are used depending on the sender of a message.