adds guards to RX and TX doneISR
In Gemini mode a noisy environment with lots of transmitters and interference can cause the radios to get a few 10us out of sync.  This can then cause the following tx done IRQ to be out of step and double call TXdoneISR.  Which can cause a double freq hop.  This is also a simple guard during normal operation.