Gemini (#1914)
* add generic gemini tx target

* add FHSS func for offset frequency

* update tx to set 2 different radio freqs

* update rx and rxLua for gemini mode

* dont use SX1280_RADIO_SET_AUTOFS in Gemini mode

Using SX1280_RADIO_SET_AUTOFS causes the 2 radios to end up in different modes e.g. FS after receiving a packet and standby after timeout.  This causes issues with dio timings.

* adds guards to RX and TX doneISR

In Gemini mode a noisy environment with lots of transmitters and interference can cause the radios to get a few 10us out of sync.  This can then cause the following tx done IRQ to be out of step and double call TXdoneISR.  Which can cause a double freq hop.  This is also a simple guard during normal operation.

* Updated tx target for testing, but this should not be used for a production tx!

* correct RXdoneISR guard

* add missing SetFrequencyReg

* update tx to be internal

* adds FHSSgetInitialGeminiFreq()

* add Gemini option to tx Lua

* add uplink_RSSI_2 and active_antenna updates

* add isDualRadio()

* update tx target for e28 modules and backpack

* Remove unnecessary boot0 high write...

Setting this gpio high was causing the pin to oscillate .  NFI why.  The write is not required and boot0 can be left low during flashing.

* add back SetFrequencyHz for startCWTest

* update sx127x lib

* change SX12XX_Radio_Number_t name

* update SetFrequencyHz to use SX12XX_Radio_Number_t

* fixes after merging in master

* update TXenable for gemini

* more cleaning after master merge

* adds LBT to gemini mode

* add power_lna_gain to gemini tx

* correct LBT for the Default single antenna state

* adds tx_all_enable_set/clr_bits

* kill all the magic numbers!

* add comment for SX1280_RADIO_SET_AUTOFS

* gemini check when calling SetRFLinkRate()...

... otherwise it isnt set correctly after a config change.

Co-authored-by: Jye <JyeSmith@users.noreply.github.com>