SNR report based on Fuzzy Logic
When the difference between those two SNRs are larger than threshold, we'd return the averaged value between two. When the difference between the two values are small, return the lower SNR value. On the borderline differences, the transition between two logics are smoothend by a fuzzy logic.

The thershold of the fuzzy logic is determined by dynamic power up/down SNR thresholds