Fix PWM channel reset on TX/RX serial port config change via LUA
Previously whenever one of the RX or TX pins (that are used for PWM
channels) was reconfigured via the LUA script (from the TX) the
corresponding sibling channel (in case of the RX pin, the TX pin and
vice versa) would also be reconfigured. That is a sane approach when
there is only serial and PWM modes available.

Fortunately ELRS allows us to change the PWM frequency of the outputs.
So an output channel can be in either serial, on/off, PWM 50Hz, PWM
100Hz, 333Hz, 400Hz or 10kHzDuty modes. Whenever any of these modes
are selected (for either of the serial pins) the entire port
configuration is also done for the sibling port. The caveat with the
previous implementation (before this commit) is that a sibling port
could only be in either serial or PWM 50Hz mode. If you wanted to
configure multiple ports for e.g. 333Hz PWM you'd not be able to
configure both of the serial ports (pin 1 & 3) for anything but 50Hz
PWM as that value was the hard coded PWM value in the serial
configuration code.

With the code path for reconfiguring a serial pin into PWM has become
a tiny bit more complex. When a port is supposed to be switched to PWM
mode but already is in PWM mode the siblings configuration will be
retained.

This now allows for sibling serial ports to have independent PWM
frequencies configured via the ELRS lua script, just as it was
possible via the webinterface already.

Fixes #2485